Fixes: NB#181838 - Localization files and CSS style are read multiple times at startup
RevBy: Armin Berres, Adrian Yanes

Details: The quite large file commonlayouts.css is parsed 3 times during startup because it is imported in the files libmeegotouchviews.css, libmeegotouchcore.css and libmeegotouchextensions.css. The patch deletes the import in libmeegotouchextensions.css, as commonlayouts.css is not used there. The shared subset of used CSS-tags has been moved to commonbaselayouts.css to improve the performance by reducing the amount of data to parse.