New: MWindowStyle
RevBy: Daniel d'Andrada
Details: MWindowStyle has been introduced to be able to define a default
background image or color when opening a window. For performance reasons
the prefix "forcex11" is required, as it indicates that the image will
get converted into X11 pixmaps no matter which graphicsystem is used in Qt.