Fixes: NB#261166 - Black line is flashed when changing the orientation in create event view
RevBy: Adrian Yanes

Details:
Since the MSheet's background is not being set with the default style
(only its header and central slots), when rotating the widget would
show a flickering black line which is ugly.

This also prevents the inner black background from being shown when
the central widget is panned (for example, when the VKB is on).