Fixes: NB#263357 - Icons fade to black on a "pressed down state".
Details: Removing the highlighter effect from list items. This means
that icons are not being dimmed anymore, but background and text changes
color as specified. If icons should be dimmed, we
need to redesign the pressed state of the whole list item, just putting an
overlay on an already dimmed background does not make sense.